+++ TestApp/ParkingApp/App.xaml.cs
@@ -1,10 +1,11 @@
 ﻿using System;
+using ParkingApp.Pages;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
 [assembly: XamlCompilation(XamlCompilationOptions.Compile)]
 
-namespace TestApp
+namespace ParkingApp
 {
     public partial class App : Application
     {
