+++ TestApp/ParkingApp/Models/DetailParking.cs
@@ -1,0 +1,23 @@
+﻿namespace ParkingApp.Models
+{
+	/// <summary>
+	/// Модель для получения детальной информации о парковке.
+	/// </summary>
+	public class DetailParking : Parking
+	{
+		/// <summary>
+		/// Координата X.
+		/// </summary>
+		public double CoordX { get; set; }
+
+		/// <summary>
+		/// Координата Y.
+		/// </summary>
+		public double CoordY { get; set; }
+
+		/// <summary>
+		/// Ссылка на камеру.
+		/// </summary>
+		public string Camera { get; set; }
+	}
+}
