+++ TestApp/ParkingApp/Pages/MainPage.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="TestApp.MainPage" Title="Список парковок">
+             x:Class="ParkingApp.Pages.MainPage" Title="Список парковок">
 	<StackLayout>
 		<ListView x:Name="ParkingList"
 			SelectedItem="{Binding SelectedParking, Mode=TwoWay}"
