+++ TestApp/ParkingApp/Pages/ParkingPage.xaml
@@ -2,7 +2,7 @@
 
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="TestApp.ParkingPage">
+             x:Class="ParkingApp.Pages.ParkingPage">
 
 	<StackLayout>
 		<StackLayout>
