+++ TestApp/ParkingApp/ParkingApp.csproj
@@ -13,4 +13,15 @@
         <PackageReference Include="System.Text.Json" Version="6.0.0-rc.2.21480.5" />
         <PackageReference Include="Xamarin.Forms" Version="5.0.0.2012" />
     </ItemGroup>
+
+    <ItemGroup>
+      <Compile Update="Pages\MainPage.xaml.cs">
+        <DependentUpon>MainPage.xaml</DependentUpon>
+        <SubType>Code</SubType>
+      </Compile>
+      <Compile Update="Pages\ParkingPage.xaml.cs">
+        <DependentUpon>ParkingPage.xaml</DependentUpon>
+        <SubType>Code</SubType>
+      </Compile>
+    </ItemGroup>
 </Project>
