+++ TestApp/TestApp.Android/MainActivity.cs
@@ -5,6 +5,7 @@ using Android.Runtime;
 using Android.Views;
 using Android.Widget;
 using Android.OS;
+using ParkingApp;
 
 namespace TestApp.Android
 {
