--- TestApp/TestApp/MainPage.xaml.cs
@@ -1,21 +1,0 @@
-﻿using Xamarin.Forms;
-
-namespace TestApp
-{
-	public partial class MainPage : ContentPage
-	{
-		private ApplicationViewModel ViewModel { get; set; }
-		public MainPage()
-		{
-			InitializeComponent();
-			ViewModel = new ApplicationViewModel {Navigation = this.Navigation};
-			this.BindingContext = ViewModel;
-		}
-
-		protected override async void OnAppearing()
-		{
-			await ViewModel.GetParkings();
-			base.OnAppearing();
-		}
-	}
-}
