--- TestApp/TestApp/Models/DetailParking.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace TestApp.Models
-{
-	public class DetailParking : Parking
-	{
-		public double CoordX { get; set; }
-		public double CoordY { get; set; }
-		public string Camera { get; set; }
-	}
-}
