--- TestApp/TestApp/Models/Parking.cs
@@ -1,10 +1,0 @@
-﻿namespace TestApp.Models
-{
-	public class Parking
-	{
-		public int Id { get; set; }
-		public string Address { get; set; }
-		public int TotalParkingSpaces { get; set; }
-		public int FreeParkingSpaces { get; set; }
-	}
-}
